New flags, but they will be released soon
Please be patient! We are adding flags and countries as fast as we can possibly do. But we release our changes too fast. In the next big update, a lot will be changed. We will add version-logs and a timer wich tells you, when the next update will come. As I said earlier, I won't leak you anything about the new flags. The release will be done on sunday, the 6th of july, so, stay tuned!

More Information in our Discord: https://dcs.lol/lokrogaming